staker memo
